Move chain_head requests to main task
